+++ src/main/java/com/skd/dataminer/init/Initializer.java
@@ -38,8 +38,6 @@ public class Initializer {
             Files.createDirectories(baseDir.resolve("performance"));
             Files.createDirectories(baseDir.resolve("performance/latency"));
             Files.createDirectories(baseDir.resolve("events/errors"));
-            Files.createDirectories(baseDir.resolve("vision/screenshots"));
-            Files.createDirectories(baseDir.resolve("vision/analyses"));
 
             generateModsJson();
             generateInfoJson();